refactor: refactored lint command
