+++ tests/integration/lint_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/kong/deck/cmd"
+	"github.com/kong/deck/lint"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"sigs.k8s.io/yaml"
@@ -29,7 +29,7 @@ func Test_LintPlain(t *testing.T) {
 	}
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			output, err := lint(
+			output, err := file_lint(
 				"-s", tc.stateFile,
 				tc.rulesetFile,
 			)
@@ -46,7 +46,7 @@ func Test_LintPlain(t *testing.T) {
 type lintErrors struct {
 	TotalCount int
 	FailCount  int
-	Results    []cmd.LintResult
+	Results    []lint.Result
 }
 
 func Test_LintStructured(t *testing.T) {
@@ -129,7 +129,7 @@ func Test_LintStructured(t *testing.T) {
 			if tc.failSeverity != "" {
 				lintOpts = append(lintOpts, "--fail-severity", tc.failSeverity)
 			}
-			output, err := lint(lintOpts...)
+			output, err := file_lint(lintOpts...)
 			require.Error(t, err)
 
 			var expectedErrors, outputErrors lintErrors
@@ -154,7 +154,7 @@ func Test_LintStructured(t *testing.T) {
 			}
 
 			cmpOpts := []cmp.Option{
-				cmpopts.SortSlices(func(a, b cmd.LintResult) bool { return a.Line < b.Line }),
+				cmpopts.SortSlices(func(a, b lint.Result) bool { return a.Line < b.Line }),
 				cmpopts.EquateEmpty(),
 			}
 			if diff := cmp.Diff(outputErrors, expectedErrors, cmpOpts...); diff != "" {
